+++ coding/coding_tests/compressed_bit_vector_test.cpp
@@ -285,7 +285,7 @@ UNIT_TEST(CompressedBitVector_Union3)
 UNIT_TEST(CompressedBitVector_Union4)
 {
   vector<uint64_t> setBits1;
-  for (int i = 0; i < coding::DenseCBV::kBlockSize; ++i)
+  for (uint64_t i = 0; i < coding::DenseCBV::kBlockSize; ++i)
     setBits1.push_back(i);
 
   vector<uint64_t> setBits2 = {1000000000};
