+++ search/features_filter.cpp
@@ -8,7 +8,7 @@
 namespace search
 {
 // FeaturesFilter ----------------------------------------------------------------------------------
-FeaturesFilter::FeaturesFilter(CBV const & filter, uint32_t threshold)
+FeaturesFilter::FeaturesFilter(CBV const & filter, uint64_t threshold)
   : m_filter(filter), m_threshold(threshold)
 {
 }
@@ -32,7 +32,7 @@ CBV LocalityFilter::Filter(CBV const & cbv) const
 }
 
 // ViewportFilter ----------------------------------------------------------------------------------
-ViewportFilter::ViewportFilter(CBV const & filter, uint32_t threshold)
+ViewportFilter::ViewportFilter(CBV const & filter, uint64_t threshold)
   : FeaturesFilter(filter, threshold)
 {
 }
